refactored shop
